New evaluation script which iterates over several setups (#106)
