+++ keys_values/__main__.py
@@ -22,6 +22,7 @@ from jsonargparse import auto_cli, set_config_read_mode, set_docstring_parse_opt
 from litgpt.__main__ import PARSER_DATA as PARSER_DATA_LITGPT
 
 from keys_values.finetune.longcontext_eval import setup as eval_long_fn
+from keys_values.finetune.longcontext_eval_ext import setup as eval_long_ext_fn
 from keys_values.finetune.longcontext_full import setup as finetune_long_full_fn
 from keys_values.finetune.longcontext_lora import setup as finetune_long_lora_fn
 from keys_values.finetune.longcon_offload_full import setup as finetune_offload_full_fn
@@ -35,9 +36,10 @@ ENV_VAR_LOG_DIR_LEGACY = "VALKEYRIE_LOG_DIR"
 
 PARSER_DATA = {
     **PARSER_DATA_LITGPT,
+    "eval_long": eval_long_fn,
+    "eval_long_ext": eval_long_ext_fn,
     "finetune_long_full": finetune_long_full_fn,
     "finetune_long_lora": finetune_long_lora_fn,
-    "eval_long": eval_long_fn,
     "finetune_offload_full": finetune_offload_full_fn,
     "finetune_offload_lora": finetune_offload_lora_fn,
 }
