+++ keys_values/scripts/remove_files_from_checkpoints.py
